main completa
