+++ main.cpp
@@ -11,7 +11,6 @@ int main(){
 
 
     std::cout<<"Digite 'cliente' caso deseje entrar como cliente"<<"\n"<<"Digite 'gerente' para gerenciar seus produtos disponiveis"<<std::endl;
-    std::cout<<"Digite 1 para cliente"<<"\n"<<"Digite 2 para Lojista"<<std::endl;
 
     std::cin>>comando;
     std::cout<<'\n'<<std::endl;
