+++ paginas.hpp
@@ -1,0 +1,13 @@
+#ifndef USUARIO_H
+#define USUARIO_H
+#include <string>
+#include <vector>
+#include <utility>
+#include <map>
+#include <iostream>
+
+void cliente_main();
+
+void loja_main();
+
+#endif
