+++ produto.hpp
@@ -14,12 +14,13 @@ public:
     friend class Estoque;
 //adicionar imagem em produto 
     Produto(std::string nome_produto, float valor, std::string descricao, std::string codigo_produto, int quantidade){
-    _nome_produto = nome_produto;
-    _valor = valor;
-    _descricao = descricao;
-    _codigo_produto = codigo_produto;
-    _quantidade = quantidade;
+        _nome_produto = nome_produto;
+        _valor = valor;
+        _descricao = descricao;
+        _codigo_produto = codigo_produto;
+        _quantidade = quantidade;
     }
+    Produto(): _valor(0.0), _quantidade(0){};
 
     void alterar_valor(float valor_novo);
     void desconto(float porcentagem);
